+++ apps/dashboard/src/components/agents/provider-dropdown.tsx
@@ -34,7 +34,8 @@ import { IS_SELF_HOSTED, IS_SELF_HOSTED_CE, SELF_HOSTED_UPGRADE_REDIRECT_URL } f
 import { useFetchIntegrations } from '@/hooks/use-fetch-integrations';
 import { useIsAgentEmailAvailable } from '@/hooks/use-is-agent-email-available';
 import { useLinkAgentIntegration } from '@/hooks/use-link-agent-integration';
-import { getAgentChannelDisplayName, getAgentChannelIconFileName } from '@/utils/agent-channel-provider-display';
+import { getAgentChannelDisplayName } from '@/utils/agent-email-provider-display';
+import { getProviderSquareIconFileName } from '@/utils/provider-square-icon';
 import { ROUTES } from '@/utils/routes';
 import { cn } from '@/utils/ui';
 import { openInNewTab } from '@/utils/url';
@@ -333,7 +334,7 @@ export function ProviderDropdown({
           <ProviderIcon
             providerId={selected.providerId}
             providerDisplayName={selected.displayName}
-            iconFileName={getAgentChannelIconFileName(selected.providerId)}
+            iconFileName={getProviderSquareIconFileName(selected.providerId)}
             className="size-4 shrink-0"
           />
           <span className="text-text-strong text-label-xs font-medium leading-4">{selected.displayName}</span>
@@ -385,7 +386,7 @@ export function ProviderDropdown({
                   <ProviderIcon
                     providerId={providerType.providerId}
                     providerDisplayName={providerType.displayName}
-                    iconFileName={getAgentChannelIconFileName(providerType.providerId)}
+                    iconFileName={getProviderSquareIconFileName(providerType.providerId)}
                     className="size-4 shrink-0"
                   />
                   <span className="text-text-sub text-label-xs min-w-0 flex-1 truncate font-medium leading-4">
@@ -540,7 +541,7 @@ export function ProviderDropdown({
                   <ProviderIcon
                     providerId={item.providerId}
                     providerDisplayName={item.displayName}
-                    iconFileName={getAgentChannelIconFileName(item.providerId)}
+                    iconFileName={getProviderSquareIconFileName(item.providerId)}
                     className="size-4 shrink-0"
                   />
                   <span className="text-text-sub text-label-xs flex-1 font-medium leading-4">{item.displayName}</span>
@@ -568,7 +569,7 @@ export function ProviderDropdown({
           <ProviderIcon
             providerId={expandedProvider.providerId}
             providerDisplayName={expandedProvider.displayName}
-            iconFileName={getAgentChannelIconFileName(expandedProvider.providerId)}
+            iconFileName={getProviderSquareIconFileName(expandedProvider.providerId)}
             className="size-4 shrink-0"
           />
           <span className="text-text-sub text-label-xs font-medium leading-4">{expandedProvider.displayName}</span>
