+++ apps/dashboard/src/components/conversations/constants.ts
@@ -1,6 +1,6 @@
 import { CONVERSATIONAL_PROVIDERS } from '@novu/shared';
 import { ConversationFiltersData } from '@/types/conversation';
-import { getAgentChannelDisplayName } from '@/utils/agent-channel-provider-display';
+import { getAgentChannelDisplayName } from '@/utils/agent-email-provider-display';
 
 export const PROVIDER_OPTIONS = CONVERSATIONAL_PROVIDERS.filter((p) => !p.comingSoon).map((p) => ({
   label: getAgentChannelDisplayName(p.providerId, p.displayName),
