+++ apps/dashboard/src/components/onboarding/connect-agent/agent-cli-success-view.tsx
@@ -6,7 +6,7 @@ import type { AgentIntegrationLink, AgentResponse } from '@/api/agents';
 import { CompletedStepIndicator } from '@/components/agents/setup-guide-primitives';
 import { ProviderIcon } from '@/components/integrations/components/provider-icon';
 import { AgentCard, type AgentCardConnectorKind } from '@/components/onboarding/claude-agent-preview-illustration';
-import { getAgentChannelDisplayName } from '@/utils/agent-channel-provider-display';
+import { getAgentChannelDisplayName } from '@/utils/agent-email-provider-display';
 
 type AgentCliSuccessViewProps = {
   agent: AgentResponse;
